fix: resolve __version__ from pyproject.toml in source checkouts (lint patch)
